WebDevelopment/staff-prof.ru/Task_2/requirements.txt
fastapi==0.115.8
fastapi-cli==0.0.7
python-multipart==0.0.20
SQLAlchemy==2.0.38
pydantic-settings==2.7.1
psycopg2-binary==2.9.10
